add rest of the options
