+++ src/datasources/asset/data-sources/calibration-forecast/CalibrationForecastDataSource.test.ts
@@ -9,9 +9,9 @@ import {
 } from "test/fixtures";
 import { SystemMetadata } from "datasources/system/types";
 import { dateTime } from "@grafana/data";
-import { AssetCalibrationForecastKey, AssetCalibrationPropertyGroupByType, AssetCalibrationTimeBasedGroupByType, AssetType, BusType, CalibrationForecastQuery, CalibrationForecastResponse, ColumnDescriptorType } from "../../types/CalibrationForecastQuery.types";
+import { AssetCalibrationForecastKey, AssetCalibrationPropertyGroupByType, AssetCalibrationTimeBasedGroupByType, CalibrationForecastQuery, CalibrationForecastResponse, ColumnDescriptorType } from "../../types/CalibrationForecastQuery.types";
 import { CalibrationForecastDataSource } from "./CalibrationForecastDataSource";
-import { AssetQueryType } from "../../types/types";
+import { AssetQueryType, AssetType, BusType } from "../../types/types";
 import { AssetCalibrationFieldNames } from "../../constants/CalibrationForecastQuery.constants";
 
 let datastore: CalibrationForecastDataSource, backendServer: MockProxy<BackendSrv>
